+++ packages/terra-overlay/CHANGELOG.md
@@ -3,13 +3,12 @@
 ## Unreleased
 
 * Changed
-  * Updated jest snapshots.
   * Updated jest snapshots for terra-icon changes.
+  * Updated jest snapshots.
 
 ## 3.69.5 - (July 5, 2022)
 
 * Changed
-
   * Minor dependency version bump
 
 ## 3.69.4 - (May 3, 2022)